requirements.txt
Django>=4.2
djangorestframework
django-filter
drf-yasg
gunicorn